Add console app for database seeding - separation of concerns Upgraded FluentValidation package
