+++ src/VotingIrregularities.sln
@@ -21,6 +21,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VotingIrregularities.Domain
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VotingIrregularities.Tests", "test\VotingIrregularities.Tests\VotingIrregularities.Tests.csproj", "{41FE87AE-428C-4CD5-88F1-A7A713334F2E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VotingIrregularities.Domain.Seed", "api\VotingIrregularities.Domain.Seed\VotingIrregularities.Domain.Seed.csproj", "{27FD17AB-3E0C-4DD2-8E2D-A537BE9366D8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -43,6 +45,10 @@ Global
 		{41FE87AE-428C-4CD5-88F1-A7A713334F2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{41FE87AE-428C-4CD5-88F1-A7A713334F2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{41FE87AE-428C-4CD5-88F1-A7A713334F2E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{27FD17AB-3E0C-4DD2-8E2D-A537BE9366D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27FD17AB-3E0C-4DD2-8E2D-A537BE9366D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27FD17AB-3E0C-4DD2-8E2D-A537BE9366D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27FD17AB-3E0C-4DD2-8E2D-A537BE9366D8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -52,6 +58,7 @@ Global
 		{73125CAC-66E8-40D7-81A7-6A847132C7B3} = {8A09B442-FB79-4293-BF9B-E34DD3AE70F3}
 		{4E70B88E-F43A-4546-955C-4A2139C1CC9C} = {388C55EB-26FF-46AB-9395-549E1A7A99AD}
 		{41FE87AE-428C-4CD5-88F1-A7A713334F2E} = {388C55EB-26FF-46AB-9395-549E1A7A99AD}
+		{27FD17AB-3E0C-4DD2-8E2D-A537BE9366D8} = {8A09B442-FB79-4293-BF9B-E34DD3AE70F3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {AF1523BC-7F31-4564-8E1B-D2DB4552FFCB}
