+++ src/api/VotingIrregularities.Api/VotingIrregularities.Api.csproj
@@ -38,7 +38,7 @@
     <PackageReference Include="SimpleInjector.Integration.AspNetCore" Version="4.3.1" />
     <PackageReference Include="SimpleInjector.Integration.AspNetCore.Mvc" Version="4.3.1" />
     <PackageReference Include="WebApiContrib.Core" Version="2.2.0" />
-    <PackageReference Include="FluentValidation" Version="7.6.104" />
+    <PackageReference Include="FluentValidation" Version="7.6.105" />
     <PackageReference Include="System.Collections.Immutable" Version="1.5.0" />
     <PackageReference Include="WindowsAzure.Storage" Version="9.3.3" />
   </ItemGroup>
