+++ src/api/VotingIrregularities.Domain.Seed/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "VotingIrregularities.Domain": {
+    "VotingIrregularities.Domain.Seed": {
       "commandName": "Project",
       "commandLineArgs": "-seed",
       "environmentVariables": {
