+++ src/test/VotingIrregularities.Domain.Tests/VotingIrregularities.Domain.Tests.csproj
@@ -18,7 +18,7 @@
     </PackageReference>
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
-    <PackageReference Include="FluentValidation" Version="7.6.104" />
+    <PackageReference Include="FluentValidation" Version="7.6.105" />
   </ItemGroup>
 
 </Project>
